200mil millones de cosas cambiadas Aañadido el include del .h donde hace falta Modificadas funciones Makefile hecho

+++ Libft/ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: rprieto- <rprieto-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/12 11:57:02 by rprieto-          #+#    #+#             */
-/*   Updated: 2019/11/13 18:02:56 by rprieto-         ###   ########.fr       */
+/*   Updated: 2019/11/21 15:20:23 by rprieto-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dst, const void *src, unsigned int n)
+#include "libft.h"
+
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	unsigned char	*dest;
 	unsigned char	*source;
