+++ Libft/ft_memmove.c
@@ -6,11 +6,13 @@
 /*   By: rprieto- <rprieto-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/13 12:10:38 by rprieto-          #+#    #+#             */
-/*   Updated: 2019/11/13 12:30:58 by rprieto-         ###   ########.fr       */
+/*   Updated: 2019/11/21 15:21:19 by rprieto-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dst, const void *src, unsigned int len)
+#include "libft.h"
+
+void	*ft_memmove(void *dst, const void *src, size_t len)
 {
 	int				i;
 	unsigned char	*source;
