+++ Libft/ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: rprieto- <rprieto-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/12 11:18:10 by rprieto-          #+#    #+#             */
-/*   Updated: 2019/11/13 09:10:30 by rprieto-         ###   ########.fr       */
+/*   Updated: 2019/11/21 15:19:09 by rprieto-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *b, int c, unsigned int len)
+#include "libft.h"
+
+void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char *pointer;
